LayoutTests/imported/w3c/web-platform-tests/content-security-policy/generic/generic-0_1-script-src-expected.txt

CONSOLE MESSAGE: Refused to load http://www1.localhost:8800/content-security-policy/generic/fail-0_1.js because it appears in neither the script-src directive nor the default-src directive of the Content Security Policy.
default-src should cascade to script-src directive


Harness Error (TIMEOUT), message = null

PASS Verify cascading of default-src to script-src policy: block
PASS Verify cascading of default-src to script-src policy: allow
NOTRUN Should fire violation events for every failed violation
